backend/requirements.txt
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.32.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
greenlet==3.1.1
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
